1- Movendo testes unitários para a classe apropriada e removendo classe de teste desnecessária no momento
